Move more gene-specifics over
