+++ context/app/app_callbacks.py
@@ -5,7 +5,8 @@ from math import log10
 
 import pandas
 import plotly.graph_objs as go
-from dash.dependencies import Input, Output, State
+from dash.dependencies import Input
+
 from plotly.figure_factory.utils import label_rgb, n_colors, unlabel_rgb
 
 from app.app_gene_callbacks import AppGeneCallbacks
@@ -38,62 +39,6 @@ class AppCallbacks(AppGeneCallbacks, AppConditionCallbacks):
         #     Output('ids-iframe', 'srcDoc'),
         #     [Input('scatter-pca', 'selectedData')]
         # )(self._update_condition_list)
-
-        self.app.callback(
-            Output('table-iframe', 'srcDoc'),
-            [Input('selected-genes-ids-json', 'children')]
-        )(self._update_gene_table)
-
-        self.app.callback(
-            Output('list-iframe', 'srcDoc'),
-            [Input('selected-genes-ids-json', 'children')]
-        )(self._update_gene_list)
-
-        # Hidden elements that record the last time a control was touched:
-
-        self.app.callback(
-            Output('search-genes-timestamp', 'children'),
-            [Input('search-genes', 'value')]
-        )(self._update_timestamp)
-
-        self.app.callback(
-            Output('scatter-sample-by-sample-timestamp', 'children'),
-            [Input('scatter-sample-by-sample', 'selectedData')]
-        )(self._update_timestamp)
-
-        self.app.callback(
-            Output('scatter-volcano-timestamp', 'children'),
-            [Input('scatter-volcano', 'selectedData')]
-        )(self._update_timestamp)
-
-        # Hidden elements which transform inputs into lists of IDs:
-
-        self.app.callback(
-            Output('search-genes-ids-json', 'children'),
-            [Input('search-genes', 'value')]
-        )(self._search_to_ids_json)
-
-        self.app.callback(
-            Output('scatter-sample-by-sample-ids-json', 'children'),
-            [Input('scatter-sample-by-sample', 'selectedData')]
-        )(self._scatter_to_ids_json)
-
-        self.app.callback(
-            Output('scatter-volcano-ids-json', 'children'),
-            [Input('scatter-volcano', 'selectedData')]
-        )(self._scatter_to_ids_json)
-
-        # Hidden elements which pick the value from the last modified control:
-
-        self.app.callback(
-            Output('selected-genes-ids-json', 'children'),
-            [Input('search-genes-timestamp', 'children'),
-             Input('scatter-sample-by-sample-timestamp', 'children'),
-             Input('scatter-volcano-timestamp', 'children')],
-            [State('search-genes-ids-json', 'children'),
-             State('scatter-sample-by-sample-ids-json', 'children'),
-             State('scatter-volcano-ids-json', 'children')]
-        )(self._pick_latest)
 
     def _search_to_ids_json(self, input):
         ids = [
@@ -185,101 +130,6 @@ class AppCallbacks(AppGeneCallbacks, AppConditionCallbacks):
     #         'layout': _ScatterLayout(x_axis, y_axis)
     #     }
 
-    def _update_scatter_genes(
-            self,
-            selected_genes_ids_json,
-            x_axis, y_axis, scale):
-        is_log = scale == 'log'
-        all = self._dataframe
-        selected = self._filter_by_genes_ids_json(
-            all,
-            selected_genes_ids_json
-        )
-        return {
-            'data': [
-                go.Scattergl(
-                    x=all[x_axis],
-                    y=all[y_axis],
-                    mode='markers',
-                    text=all.index,
-                    marker=_light_dot,
-                ),
-                go.Scattergl(
-                    x=selected[x_axis],
-                    y=selected[y_axis],
-                    mode='markers',
-                    text=selected.index,
-                    marker=_dark_dot
-                )
-            ],
-            'layout': _ScatterLayout(
-                x_axis, y_axis,
-                x_log=is_log, y_log=is_log)
-        }
-
-    def _update_scatter_volcano(
-            self,
-            selected_genes_ids_json,
-            file_selected,
-            x_axis, y_axis):
-        if not x_axis:
-            # ie, there are no differential files.
-            # "file" itself is (mis)used for messaging.
-            return {}
-        all = self._diff_dataframes[file_selected]
-        selected = self._filter_by_genes_ids_json(
-            all,
-            selected_genes_ids_json
-        )
-        return {
-            'data': [
-                go.Scattergl(
-                    x=all[x_axis],
-                    y=all[y_axis],
-                    mode='markers',
-                    text=self._dataframe.index,
-                    marker=_light_dot
-                ),
-                go.Scattergl(
-                    x=selected[x_axis],
-                    y=selected[y_axis],
-                    mode='markers',
-                    text=self._dataframe.index,
-                    marker=_dark_dot
-                )
-            ],
-            'layout': _ScatterLayout(x_axis, y_axis)
-        }
-    #
-    # def _update_condition_list(self, selected_data):
-    #     points = [point['pointNumber'] for point in selected_data['points']]
-    #     return self._list_html(self._dataframe.T.iloc[points])
-    #     # Alternatively:
-    #     #   pandas.DataFrame(self._dataframe.columns.tolist())
-    #     # but transpose may be more efficient than creating a new DataFrame.
-
-    def _update_gene_table(self, selected_genes_ids_json):
-        selected_genes_df = self._filter_by_genes_ids_json(
-            self._dataframe,
-            selected_genes_ids_json
-        )
-        return self._table_html(selected_genes_df)
-
-    def _update_gene_list(self, selected_genes_ids_json):
-        selected_genes_df = self._filter_by_genes_ids_json(
-            self._dataframe,
-            selected_genes_ids_json
-        )
-        return self._list_html(selected_genes_df)
-
-    def _filter_by_genes_ids_json(self, dataframe, json_list):
-        selected_genes_ids = json.loads(json_list)
-        selected_genes = [
-            item for (i, item)
-            in enumerate(self._genes)
-            if i in selected_genes_ids
-        ]
-        return dataframe.loc[selected_genes]
 
     def _table_html(self, dataframe):
         """
@@ -311,14 +161,7 @@ class AppCallbacks(AppGeneCallbacks, AppConditionCallbacks):
         ])
 
 
-_dark_dot = {
-    'color': 'rgb(0,0,127)',
-    'size': 5
-}
-_light_dot = {
-    'color': 'rgb(127,216,127)',
-    'size': 5
-}
+
 
 
 def _remove_rowname_header(s):
@@ -374,23 +217,4 @@ def _match_booleans(search_term, index_set, targets):
     ]
 
 
-class _ScatterLayout(go.Layout):
-    def __init__(self, x_axis, y_axis, x_log=False, y_log=False):
-        x_axis_config = {'title': x_axis}
-        y_axis_config = {'title': y_axis}
-        if x_log:
-            x_axis_config['type'] = 'log'
-        if y_log:
-            y_axis_config['type'] = 'log'
-        super().__init__(
-            showlegend=False,
-            xaxis=x_axis_config,
-            yaxis=y_axis_config,
-            margin=go.Margin(
-                l=80,  # noqa: E741
-                r=20,
-                b=60,
-                t=20,
-                pad=5
-            )
-        )
+
