+++ context/app/utils/callbacks.py
@@ -1,5 +1,8 @@
 from dash.dependencies import Output, Input
 
+
+import plotly.graph_objs as go
+
 def figure_output(id):
     return Output(id, 'figure')
 
@@ -14,3 +17,33 @@ def scatter_inputs(id, scale_select=False):
             Input('scale-select', 'value')
         )
     return inputs
+
+dark_dot = {
+    'color': 'rgb(0,0,127)',
+    'size': 5
+}
+light_dot = {
+    'color': 'rgb(127,216,127)',
+    'size': 5
+}
+
+class ScatterLayout(go.Layout):
+    def __init__(self, x_axis, y_axis, x_log=False, y_log=False):
+        x_axis_config = {'title': x_axis}
+        y_axis_config = {'title': y_axis}
+        if x_log:
+            x_axis_config['type'] = 'log'
+        if y_log:
+            y_axis_config['type'] = 'log'
+        super().__init__(
+            showlegend=False,
+            xaxis=x_axis_config,
+            yaxis=y_axis_config,
+            margin=go.Margin(
+                l=80,  # noqa: E741
+                r=20,
+                b=60,
+                t=20,
+                pad=5
+            )
+        )
